feat(engine): Add Sized2D trait + move Size to layout module
