+++ mindy-engine/src/mindmap/layout.rs
@@ -2,7 +2,7 @@ use crate::mindmap::Mindmap;
 use crate::mindmap::style::MindmapStyle;
 use crate::node::Node;
 use crate::layout::pos2::Pos2;
-use crate::utils::size::Size;
+use crate::layout::size::Size;
 
 pub struct LayoutEngine {}
