+++ mindy-engine/src/utils/mod.rs
@@ -1,3 +1,2 @@
 pub mod rgb;
-pub mod size;
 pub mod throttler;
