--- mindy-engine/src/utils/size.rs
@@ -1,13 +1,0 @@
-use serde::{Deserialize, Serialize};
-
-#[derive(Debug, Clone, Default, PartialEq, Deserialize, Serialize)]
-pub struct Size {
-    pub width: f32,
-    pub height: f32,
-}
-
-impl Size {
-    pub fn new(width: f32, height: f32) -> Self {
-        Self { width, height }
-    }
-}
